# swagger.yaml
openapi: 3.0.3
info:
  title: User Management
  description: |-
    Project Requirements
      - Build a digital newspaper system
        - Scope 2 authority: ADMIN, USER
        - Admin/user can login/logout, update their profile, password
        - User basic information: email, phone, firstname, lastname, dob
        - Register only for user
        - Admin can view, edit, and delete user; but user can’t
  contact:
    email: ndphuong-batch43@sdc.edu.vn
  version: 1.0.11
servers:
  - url: http://localhost:8000/api
...
paths:
...
  /auth/createUser:
    post:
      tags:
        - auth
      summary: Create user
      description: Create user to server
      operationId: createUser
      security:
        - BearerAuth: []
      requestBody:
        description: Create user object
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
        required: true
      responses:
        "200":
          description: user Created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    type: object
                    properties:
                      ok:
                        type: boolean
                        example: "true"
                      message:
                        type: string
                        example: "User Created successfully!!!"
                  resData:
                    $ref: "#/components/schemas/Logged"
        "400":
          description: Validation error in request body/user creation failed
...
components:
  schemas:
    User:
      type: object
      properties:
        firstName:
          type: string
          example: sin
        lastName:
          type: string
          example: lee
        email:
          type: string
          example: leesin@gmail.com
        phone:
          type: string
          example: 0362598631
        password:
          type: string
          example: leesin123
        dob:
          type: string
          example: 1999-01-01
...
    Logged:
      type: object
      properties:
        accessToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjM0MzJlMTE5LTMyNWQtNDcwNi05NmUxLThhMWQ4ZjNiMTU0OSIsInJvbGUiOiJ1c2VyIiwiaWF0IjoxNzA5MTEwMDI2LCJleHAiOjE3MDkxMTM2MjZ9.DZsR_JZj97AdM3agay7DPIPaLyYZBFdCoRXo0JTm0DU
        userData:
          $ref: "#/components/schemas/User"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT